update cmakelists for cuda, replace double by float

+++ include/filter.hpp
@@ -30,23 +30,23 @@ namespace kalman
        * @param matchCount The number of time this observation was matched
        * @param t The position at which the observation was made
        */
-      Observation(kMatrix<double> obs, uint32_t matchCount, uint32_t t)
+      Observation(kMatrix<float> obs, uint32_t matchCount, uint32_t t)
               : obs(std::move(obs))
               , match_count(matchCount)
               , t(t)
       {
       }
-//    Eigen::Matrix<double, 3, 1> obs;         // The observation Matrix
-      kMatrix<double> obs;         // The observation Matrix
+//    Eigen::Matrix<float, 3, 1> obs;         // The observation Matrix
+      kMatrix<float> obs;         // The observation Matrix
       uint32_t                    match_count; // The numbers of time the observation was matched to a filter
       uint32_t                    t;           // The t-position at which the observation was made
   };
 
-    static const kMatrix<double> A({1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}, 4, 4);
-    static const kMatrix<double> A_transpose({1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}, 4, 4);
-    static const kMatrix<double> C({1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}, 3, 4);
-    static const kMatrix<double> C_transpose({1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1}, 4, 3);
-    static const kMatrix<double> Vn({SEGDET_VARIANCE_POSITION, 0, 0, 0,
+    static const kMatrix<float> A({1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}, 4, 4);
+    static const kMatrix<float> A_transpose({1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}, 4, 4);
+    static const kMatrix<float> C({1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1}, 3, 4);
+    static const kMatrix<float> C_transpose({1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1}, 4, 3);
+    static const kMatrix<float> Vn({SEGDET_VARIANCE_POSITION, 0, 0, 0,
                                      SEGDET_VARIANCE_THICKNESS, 0, 0, 0, SEGDET_VARIANCE_LUMINOSITY}, 3, 3);
 
 
@@ -62,33 +62,33 @@ namespace kalman
          * @param slopeMax The maximum value that the filter's slope can have
          * @param observation The filter's first observation
          */
-//    Filter(bool isHorizontal, uint32_t t_integration, double slopeMax, Eigen::Matrix<double, 3, 1> observation)
-        Filter(bool isHorizontal, uint32_t t_integration, double slopeMax, kMatrix<double> observation)
+//    Filter(bool isHorizontal, uint32_t t_integration, float slopeMax, Eigen::Matrix<float, 3, 1> observation)
+        Filter(bool isHorizontal, uint32_t t_integration, float slopeMax, kMatrix<float> observation)
                 : is_horizontal(isHorizontal)
                 , slope_max(slopeMax)
-                , S(kMatrix<double>({observation(0, 0), 0, observation(1, 0), observation(2, 0)}, 4, 1))
-                , W(kMatrix<double>({0, 0, 0, 0}, 4, 1))
-//      , W(Eigen::Matrix<double, 4, 1>::Zero())
-                , N(kMatrix<double>({0,0,0}, 3, 1))
-//      , N(Eigen::Matrix<double, 3, 1>::Zero())
-                , H(kMatrix<double>({1, 0, 0, 0,
+                , S(kMatrix<float>({observation(0, 0), 0, observation(1, 0), observation(2, 0)}, 4, 1))
+                , W(kMatrix<float>({0, 0, 0, 0}, 4, 1))
+//      , W(Eigen::Matrix<float, 4, 1>::Zero())
+                , N(kMatrix<float>({0,0,0}, 3, 1))
+//      , N(Eigen::Matrix<float, 3, 1>::Zero())
+                , H(kMatrix<float>({1, 0, 0, 0,
                                      0, 1, 0, 0,
                                      0, 0, 1, 0 ,
                                      0, 0, 0,1}, 4, 4))
-//      , H(Eigen::Matrix<double, 4, 4>::Identity())
+//      , H(Eigen::Matrix<float, 4, 4>::Identity())
                 , observation(std::nullopt)
                 , observation_distance(0)
                 , last_integration(t_integration)
-                , S_predicted(kMatrix<double>({0, 0,0, 0}, 4, 1))
-                , X_predicted(kMatrix<double>({0, 0, 0}, 3, 1))
+                , S_predicted(kMatrix<float>({0, 0,0, 0}, 4, 1))
+                , X_predicted(kMatrix<float>({0, 0, 0}, 3, 1))
         {
             first    = t_integration;
             t_values = std::vector<uint32_t>({t_integration});
             n_values = std::vector<uint32_t>({static_cast<uint32_t>(observation(0, 0))});
 
-            slopes                           = std::vector<double>({0});
-            thicknesses                      = std::vector<double>({observation(1, 0)});
-            luminosities                     = std::vector<double>({observation(2, 0)});
+            slopes                           = std::vector<float>({0});
+            thicknesses                      = std::vector<float>({observation(1, 0)});
+            luminosities                     = std::vector<float>({observation(2, 0)});
             nb_current_slopes_over_slope_max = 0;
             first_slope                      = std::nullopt;
 
@@ -103,8 +103,8 @@ namespace kalman
             sigma_thickness  = SEGDET_DEFAULT_SIGMA_THK;
             sigma_luminosity = SEGDET_DEFAULT_SIGMA_LUM;
 
-//      S_predicted = Eigen::Matrix<double, 4, 1>::Zero();
-//      X_predicted = Eigen::Matrix<double, 3, 1>::Zero();
+//      S_predicted = Eigen::Matrix<float, 4, 1>::Zero();
+//      X_predicted = Eigen::Matrix<float, 3, 1>::Zero();
 
         }
 
@@ -115,35 +115,35 @@ namespace kalman
         std::vector<uint32_t> n_values;
 
         uint                  nb_current_slopes_over_slope_max;
-        double                slope_max;
-        std::optional<double> first_slope;
-        std::vector<double>   thicknesses;
-        std::vector<double>   luminosities;
-        std::vector<double>   slopes;
+        float                slope_max;
+        std::optional<float> first_slope;
+        std::vector<float>   thicknesses;
+        std::vector<float>   luminosities;
+        std::vector<float>   slopes;
 
         std::vector<Point> under_other; //
         std::vector<Point> currently_under_other;
         std::vector<Point> segment_points;
 
-//    Eigen::Matrix<double, 4, 1> S; // state matrix {{position (n)}, {slope}, {thickness}, {luminosity}}
-//    Eigen::Matrix<double, 4, 1> W; // noise matrix
-//    Eigen::Matrix<double, 3, 1> N; // measured noise matrix
-//    Eigen::Matrix<double, 4, 4> H; // S prediction error variance matrix
-        kMatrix<double> S; // state matrix {{position (n)}, {slope}, {thickness}, {luminosity}}
-        kMatrix<double> W; // noise matrix
-        kMatrix<double> N; // measured noise matrix
-        kMatrix<double> H; // S prediction error variance matrix
+//    Eigen::Matrix<float, 4, 1> S; // state matrix {{position (n)}, {slope}, {thickness}, {luminosity}}
+//    Eigen::Matrix<float, 4, 1> W; // noise matrix
+//    Eigen::Matrix<float, 3, 1> N; // measured noise matrix
+//    Eigen::Matrix<float, 4, 4> H; // S prediction error variance matrix
+        kMatrix<float> S; // state matrix {{position (n)}, {slope}, {thickness}, {luminosity}}
+        kMatrix<float> W; // noise matrix
+        kMatrix<float> N; // measured noise matrix
+        kMatrix<float> H; // S prediction error variance matrix
 
-        double n_min;
-        double n_max;
-        double sigma_position;
-        double sigma_thickness;
-        double sigma_luminosity;
+        float n_min;
+        float n_max;
+        float sigma_position;
+        float sigma_thickness;
+        float sigma_luminosity;
 
-//    Eigen::Matrix<double, 4, 1> S_predicted;
-//    Eigen::Matrix<double, 3, 1> X_predicted;
-        kMatrix<double> S_predicted;
-        kMatrix<double> X_predicted;
+//    Eigen::Matrix<float, 4, 1> S_predicted;
+//    Eigen::Matrix<float, 3, 1> X_predicted;
+        kMatrix<float> S_predicted;
+        kMatrix<float> X_predicted;
 
         std::optional<Observation>
                 observation;          // matrix {{position (n)},{thickness},{luminosity}}, nullopt if none was matched
@@ -173,7 +173,7 @@ namespace kalman
      * @param max The observation max value
      * @return true if observation is compatible with filter else false
      */
-    bool accepts(const Filter& filter, const kMatrix<double>& obs, uint32_t min, uint32_t max,
+    bool accepts(const Filter& filter, const kMatrix<float>& obs, uint32_t min, uint32_t max,
                  Parameters params);
 
   /**
