+++ include/observation_parser.hh
@@ -5,10 +5,10 @@
 namespace kalman {
     class obs_parser {
         public:
-            std::vector<std::vector<kMatrix<double>>> parse(int width, int height, std::vector<u_int8_t> img, int threshold) {
-                std::vector<std::vector<kMatrix<double>>> vec;
+            std::vector<std::vector<kMatrix<float>>> parse(int width, int height, std::vector<u_int8_t> img, int threshold) {
+                std::vector<std::vector<kMatrix<float>>> vec;
                 for(int j = 0; j < width; j++){
-                    std::vector<kMatrix<double>> tmp_vec;
+                    std::vector<kMatrix<float>> tmp_vec;
                     auto max = -1;
                     //auto pos_max = -1;
                     auto start = -1;
@@ -25,9 +25,9 @@ namespace kalman {
                         }
                         else {
                             if (max != -1) {
-                                tmp_vec.push_back(kMatrix<double>({static_cast<double>((start + i) / 2),
-                                                   static_cast<double>(i - start),
-                                                   static_cast<double>(max)}, 3, 1));
+                                tmp_vec.push_back(kMatrix<float>({static_cast<float>((start + i) / 2),
+                                                   static_cast<float>(i - start),
+                                                   static_cast<float>(max)}, 3, 1));
                             }
                             max = -1;
                             //pos_max = -1;
@@ -35,9 +35,9 @@ namespace kalman {
                         }
                     }
                     if (max != -1)
-                        tmp_vec.push_back(kMatrix<double>({static_cast<double>((start + height) / 2),
-                                           static_cast<double>(height - start),
-                                           static_cast<double>(max)}, 3, 1));
+                        tmp_vec.push_back(kMatrix<float>({static_cast<float>((start + height) / 2),
+                                           static_cast<float>(height - start),
+                                           static_cast<float>(max)}, 3, 1));
                     vec.push_back(tmp_vec);
                 }
                 return vec;
